+++ src/main/java/zaftnotameni/creatania/recipes/cobblegen/AllCobblegenRecipes.java
@@ -1,20 +1,20 @@
 package zaftnotameni.creatania.recipes.cobblegen;
+
+import java.util.ArrayList;
+import java.util.List;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.registration.IRecipeRegistration;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
-
-import java.util.ArrayList;
-import java.util.List;
 public class AllCobblegenRecipes {
   public static List<CobblegenRecipe> all = new ArrayList<>();
   public static void populateRecipes(ServerLevel serverLevel) {
     all = serverLevel.getRecipeManager().getAllRecipesFor(CobblegenRecipe.TYPE);
   }
   public static void populateRecipes(ClientLevel clientLevel) {
-    all = clientLevel.getRecipeManager().getAllRecipesFor(CobblegenRecipe.TYPE);
+    all = clientLevel.getRecipeManager().getAllRecipesFor(CobblegenRecipe.TYPE).stream().filter(CobblegenRecipe::isValid).toList();
   }
   public static List<CobblegenRecipe> getCobblegenRecipes(LevelAccessor level) {
     if ((all == null || all.isEmpty()) && level instanceof ClientLevel clientLevel) populateRecipes(clientLevel);